# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for AppStoreConnect project",
        "title": "AppStoreConnect project API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/{tenantId}/{integrationId}",
    "paths": {
...
        "/profiles": {
            "get": {
                "description": "Get all profiles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get profiles list",
                "operationId": "get-profiles-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/entities.Profile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new profile instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "(NOT IMPLEMENTED) Add new profile",
                "operationId": "add-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "profile params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Profile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/status/{id}": {
            "post": {
                "description": "Get command status by command id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Check status",
                "operationId": "get-status-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "command identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/{id}": {
            "get": {
                "description": "Get profile instance by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get profile",
                "operationId": "get-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "profile identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Profile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/pkg_httpserver_handlers.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Profile deletion by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "(NOT IMPLEMENTED) Delete profile",
                "operationId": "delete-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenantId",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "integrationId",
                        "name": "integrationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "profile identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deletion success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.Profile": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
        "pkg_httpserver_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}